+++ frontend/webui/src/App.vue
@@ -19,6 +19,7 @@ import ImageViewerModal from './components/ui/ImageViewerModal.vue';
 import ConnectionOverlay from './components/ui/ConnectionOverlay.vue';
 import MaintenanceOverlay from './components/ui/MaintenanceOverlay.vue';
 
+
 // Async Modals
 const SlideshowModal = defineAsyncComponent(() => import('./components/modals/SlideshowModal.vue'));
 const CommandOutputModal = defineAsyncComponent(() => import('./components/modals/CommandOutputModal.vue'));