+++ frontend/webui/src/components/modals/ModalContainer.vue
@@ -69,7 +69,8 @@ const ScrapeUrlModal = defineAsyncComponent(() => import('./ScrapeUrlModal.vue')
 const ArtefactImportWizardModal = defineAsyncComponent(() => import('./ArtefactImportWizardModal.vue'));
 // Registered the NotebookWizardModal here
 const NotebookWizardModal = defineAsyncComponent(() => import('./NotebookWizardModal.vue'));
-
+const AdminCreateUserModal = defineAsyncComponent(() => import('./AdminCreateUserModal.vue'));
+const CreateArtefactModal = defineAsyncComponent(() => import('./CreateArtefactModal.vue'));
 </script>
 
 <template>
@@ -137,5 +138,8 @@ const NotebookWizardModal = defineAsyncComponent(() => import('./NotebookWizardM
     <ArtefactImportWizardModal v-if="activeModal === 'artefactImportWizard'" />
     <!-- Added the modal here -->
     <NotebookWizardModal v-if="activeModal === 'notebookWizard'" />
+    <AdminCreateUserModal v-if="activeModal === 'adminCreateUser'" />
+    <CreateArtefactModal v-if="activeModal === 'createArtefact'" />
+    
   </div>
 </template>